Project Plan + Design Pattern explanation added. Also changes to the project.
 - Added the AmmoFacade Class to easily implement more Objects of the very
   complex Usable Class when "ammo_mod" behaviour is wanted
 - Minor change to DamageTrait.cs (now takes an int)

+++ ConsoleDungeonCrawler/ConsoleDungeonCrawler/Traits/DamageTrait.cs
@@ -6,10 +6,10 @@ using System.Threading.Tasks;
 
 public class DamageTrait : ITraitBehaviour
 {
-    private float dmg;
+    private int dmg;
 
 
-    public DamageTrait(float dmg)
+    public DamageTrait(int dmg)
     {
         this.dmg = dmg;
     }
